refactor: ReservationRepository 클래스를 생성해 reservations 생성 및 보관 책임 이동
